+++ src/buttons/panel_hub_refresh.js
@@ -1,0 +1,9 @@
+const { createPanelHubMessage } = require("../utils/panelHub");
+
+module.exports = {
+  customId: "panel_hub_refresh",
+
+  async execute(interaction) {
+    return interaction.update(createPanelHubMessage());
+  }
+};